dev-requirements.txt
-e git+https://github.com/cloudify-incubator/cloudify-plugins-tools.git@0.1.4#egg=cloudify-plugins-tools

cloudify-common>=5.0.0
etcd3
